Invoice Service
Added a basic outline of the invoice service. The outline contains one Resource providing a get request to get all invoices, a main file and a unit test to test that the get request to the resource is working as intended.
The commit also adds a Dockerfile (without content for now) and a requirements.txt file to include all python dependencies.